+++ Assets/Prefabs/Mushrooms/mushroom3_final.prefab
@@ -84,7 +84,7 @@ Transform:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1868099317067230}
   m_LocalRotation: {x: -0.6851938, y: -0.25556287, z: 0.36586955, w: 0.57561845}
-  m_LocalPosition: {x: -1.72, y: 1.006, z: 0.2}
+  m_LocalPosition: {x: -1.72, y: 1.406, z: 0.2}
   m_LocalScale: {x: 9.26, y: 9.26, z: 9.26}
   m_Children: []
   m_Father: {fileID: 4464928900873914}
@@ -113,7 +113,7 @@ Transform:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1492291032566892}
   m_LocalRotation: {x: -0.58621806, y: -0.3747095, z: 0.39540917, w: 0.5996606}
-  m_LocalPosition: {x: -2, y: 1.3, z: 0.14}
+  m_LocalPosition: {x: -2, y: 1.7, z: 0.14}
   m_LocalScale: {x: 38.75, y: 38.75, z: 38.75}
   m_Children: []
   m_Father: {fileID: 4464928900873914}
@@ -126,7 +126,7 @@ Transform:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1886014802587162}
   m_LocalRotation: {x: -0.7071068, y: 0, z: 0, w: 0.7071068}
-  m_LocalPosition: {x: -0, y: -1.67, z: 0}
+  m_LocalPosition: {x: -0, y: -1.27, z: 0}
   m_LocalScale: {x: 40.77, y: 40.77, z: 40.77}
   m_Children: []
   m_Father: {fileID: 4464928900873914}
